+++ R/divide_method.R
@@ -9,7 +9,7 @@ setGeneric("divide", function(.Object, n) standardGeneric("divide"))
 #' @exportMethod divide
 setMethod("divide", "matrix", function(.Object, n){
   chunkFactor <- cut(
-    c(1:nrow(.Object)),
+    1:nrow(.Object),
     unique(c(1, floor(c(1:(n-1))*(nrow(.Object)/n)), nrow(.Object))),
     include.lowest=TRUE
   )
@@ -20,7 +20,7 @@ setMethod("divide", "matrix", function(.Object, n){
 #' @rdname divide
 setMethod("divide", "vector", function(.Object, n){
   chunkFactor <- cut(
-    c(1:length(.Object)),
+    1:length(.Object),
     unique(c(1, floor(c(1:(n-1))*(length(.Object)/n)), length(.Object))),
     include.lowest=TRUE
   )